refactor(utils): move notify module to utils directory
Relocate the notify module from the root CopilotChat directory to
lua/CopilotChat/utils/notify.lua for better organization and consistency.
Update all require statements to use the new path.

No functional changes; this is a structural refactor to improve codebase
maintainability.
