+++ lua/CopilotChat/config/providers.lua
@@ -1,7 +1,7 @@
 local log = require('plenary.log')
 local plenary_utils = require('plenary.async.util')
 local constants = require('CopilotChat.constants')
-local notify = require('CopilotChat.notify')
+local notify = require('CopilotChat.utils.notify')
 local utils = require('CopilotChat.utils')
 local curl = require('CopilotChat.utils.curl')
 local files = require('CopilotChat.utils.files')
