+++ lua/CopilotChat/prompts.lua
@@ -1,7 +1,7 @@
 local client = require('CopilotChat.client')
 local constants = require('CopilotChat.constants')
 local functions = require('CopilotChat.functions')
-local notify = require('CopilotChat.notify')
+local notify = require('CopilotChat.utils.notify')
 local files = require('CopilotChat.utils.files')
 local orderedmap = require('CopilotChat.utils.orderedmap')
 local utils = require('CopilotChat.utils')
