+++ lua/CopilotChat/ui/chat.lua
@@ -1,7 +1,7 @@
 local Overlay = require('CopilotChat.ui.overlay')
 local Spinner = require('CopilotChat.ui.spinner')
 local constants = require('CopilotChat.constants')
-local notify = require('CopilotChat.notify')
+local notify = require('CopilotChat.utils.notify')
 local utils = require('CopilotChat.utils')
 local class = require('CopilotChat.utils.class')
 local orderedmap = require('CopilotChat.utils.orderedmap')
